+++ cmd/hidden_lake/messenger/pkg/app/init_service.go
@@ -35,8 +35,17 @@ func (p *sApp) initInterfaceServiceHTTP() {
 
 	msgSize := p.fWrapper.GetConfig().GetMessageSizeBytes()
 	keySize := p.fWrapper.GetConfig().GetKeySizeBits()
-	msgLimitBytes := pkg_client.GetMessageLimit(msgSize, keySize)
-	msgLimitBase64 := msgLimitBytes - (msgLimitBytes / 4) // https://ru.wikipedia.org/wiki/Base64
+	msgLimit := pkg_client.GetMessageLimit(msgSize, keySize)
+
+	// also sent encrypted session key with asymmetric encryption into hex encode
+	overHeadBytes := (2 * (keySize / 8)) + uint64(hlm_settings.CSeparatorLen)
+	msgLimitBytes := msgLimit - overHeadBytes
+	if msgLimitBytes > msgLimit { // uint64 overflow
+		panic("overflow: msgLimitBytes > msgLimit")
+	}
+
+	// overhead base64 format: https://ru.wikipedia.org/wiki/Base64
+	msgLimitBase64 := msgLimitBytes - (msgLimitBytes / 4)
 
 	mux.HandleFunc(hlm_settings.CHandleIndexPath, handler.IndexPage(p.fStateManager, p.fLogger))                                 // GET
 	mux.HandleFunc(hlm_settings.CHandleSignOutPath, handler.SignOutPage(p.fStateManager, p.fLogger))                             // GET